+++ src/browser.ts
@@ -1,9 +1,10 @@
 /* tslint:disable:no-var-requires */
+/* eslint import/no-import-module-exports: 0 */
+
+import axios from "axios"; // idk why axios is weird
 
 export * from "./index";
 export * as StellarBase from "@stellar/stellar-base";
-
-import axios from "axios"; // idk why axios is weird
 export { axios };
 
 export default module.exports;