+++ src/contract/utils.ts
@@ -1,10 +1,7 @@
-import { xdr, cereal } from "@stellar/stellar-base";
-import type { AssembledTransaction } from "./assembled_transaction";
-
-/**
- * The default timeout for waiting for a transaction to be included in a block.
- */
-export const DEFAULT_TIMEOUT = 5 * 60;
+import { xdr, cereal, Account } from "@stellar/stellar-base";
+import { Server } from "../rpc/server";
+import { type AssembledTransaction } from "./assembled_transaction";
+import { NULL_ACCOUNT , AssembledTransactionOptions } from "./types";
 
 /**
  * Keep calling a `fn` for `timeoutInSeconds` seconds, if `keepWaitingIf` is
@@ -106,4 +103,14 @@ export function processSpecEntryStream(buffer: Buffer) {
     res.push(xdr.ScSpecEntry.read(reader));
   }
   return res;
+}
+
+//eslint-disable-next-line require-await
+export async function getAccount<T>(
+  options: AssembledTransactionOptions<T>,
+  server: Server
+): Promise<Account> {
+  return options.publicKey
+    ? server.getAccount(options.publicKey)
+    : new Account(NULL_ACCOUNT, "0");
 }