+++ src/horizon/call_builder.ts
@@ -6,6 +6,7 @@ import { BadRequestError, NetworkError, NotFoundError } from "../errors";
 import { HorizonApi } from "./horizon_api";
 import { AxiosClient, version } from "./horizon_axios_client";
 import { ServerApi } from "./server_api";
+import type { Server } from "../federation";
 
 // Resources which can be included in the Horizon response via the `join`
 // query-param.
@@ -49,8 +50,11 @@ export class CallBuilder<
     | ServerApi.CollectionPage<HorizonApi.BaseResponse>
 > {
   protected url: URI;
+
   public filter: string[][];
+
   protected originalSegments: string[];
+
   protected neighborRoot: string;
 
   constructor(serverUrl: URI, neighborRoot: string = "") {
@@ -93,10 +97,10 @@ export class CallBuilder<
    * @see [Horizon Response Format](https://developers.stellar.org/api/introduction/response-format/)
    * @see [MDN EventSource](https://developer.mozilla.org/en-US/docs/Web/API/EventSource)
    * @param {object} [options] EventSource options.
-   * @param {function} [options.onmessage] Callback function to handle incoming messages.
-   * @param {function} [options.onerror] Callback function to handle errors.
+   * @param {Function} [options.onmessage] Callback function to handle incoming messages.
+   * @param {Function} [options.onerror] Callback function to handle errors.
    * @param {number} [options.reconnectTimeout] Custom stream connection timeout in ms, default is 15 seconds.
-   * @returns {function} Close function. Run to close the connection and stop listening for new events.
+   * @returns {Function} Close function. Run to close the connection and stop listening for new events.
    */
   public stream(options: EventSourceOptions<T> = {}): () => void {
     this.checkFilter();
@@ -115,11 +119,12 @@ export class CallBuilder<
     const createTimeout = () => {
       timeout = setTimeout(() => {
         es?.close();
+        // eslint-disable-next-line @typescript-eslint/no-use-before-define
         es = createEventSource();
       }, options.reconnectTimeout || 15 * 1000);
     };
 
-    let createEventSource = (): EventSource => {
+    const createEventSource = (): EventSource => {
       try {
         if (!EventSource) {
           console.warn(
@@ -192,6 +197,8 @@ export class CallBuilder<
       return es;
     };
 
+  
+
     createEventSource();
     return () => {
       clearTimeout(timeout);
@@ -213,7 +220,7 @@ export class CallBuilder<
   /**
    * Sets `limit` parameter for the current call. Returns the CallBuilder object on which this method has been called.
    * @see [Paging](https://developers.stellar.org/api/introduction/pagination/)
-   * @param {number} number Number of records the server should return.
+   * @param {number} recordsNumber Number of records the server should return.
    * @returns {object} current CallBuilder instance
    */
   public limit(recordsNumber: number): this {
@@ -239,7 +246,7 @@ export class CallBuilder<
    * will include a `transaction` field for each operation in the
    * response.
    *
-   * @param {"transactions"} join Records to be included in the response.
+   * @param "include" join Records to be included in the response.
    * @returns {object} current CallBuilder instance.
    */
   public join(include: "transactions"): this {
@@ -290,9 +297,9 @@ export class CallBuilder<
    * Convert a link object to a function that fetches that link.
    * @private
    * @param {object} link A link object
-   * @param {bool} link.href the URI of the link
-   * @param {bool} [link.templated] Whether the link is templated
-   * @returns {function} A function that requests the link
+   * @param {boolean} link.href the URI of the link
+   * @param {boolean} [link.templated] Whether the link is templated
+   * @returns {Function} A function that requests the link
    */
   private _requestFnForLink(link: HorizonApi.ResponseLink): (opts?: any) => any {
     return async (opts: any = {}) => {
@@ -321,7 +328,7 @@ export class CallBuilder<
     if (!json._links) {
       return json;
     }
-    for (const key of Object.keys(json._links)) {
+    Object.keys(json._links).forEach((key) => {
       const n = json._links[key];
       let included = false;
       // If the key with the link name already exists, create a copy
@@ -341,14 +348,16 @@ export class CallBuilder<
         const record = this._parseRecord(json[key]);
         // Maintain a promise based API so the behavior is the same whether you
         // are loading from the server or in-memory (via join).
+        // eslint-disable-next-line require-await
         json[key] = async () => record;
       } else {
         json[key] = this._requestFnForLink(n as HorizonApi.ResponseLink);
       }
-    }
+    });
     return json;
   }
 
+  // eslint-disable-next-line require-await
   private async _sendNormalRequest(initialUrl: URI) {
     let url = initialUrl;
 
@@ -404,6 +413,7 @@ export class CallBuilder<
    * @param {object} error Network error object
    * @returns {Promise<Error>} Promise that rejects with a human-readable error
    */
+  // eslint-disable-next-line require-await
   private async _handleNetworkError(error: NetworkError): Promise<void> {
     if (error.response && error.response.status && error.response.statusText) {
       switch (error.response.status) {