+++ src/horizon/strict_send_path_call_builder.ts
@@ -22,7 +22,7 @@ import { ServerApi } from "./server_api";
  *
  * Do not create this object directly, use {@link Server#strictSendPaths}.
  * @see [Find Payment Paths](https://developers.stellar.org/api/aggregations/paths/)
- * @extends CallBuilder
+ * @augments CallBuilder
  * @param {string} serverUrl Horizon server URL.
  * @param {Asset} sourceAsset The asset to be sent.
  * @param {string} sourceAmount The amount, denominated in the source asset, that any returned path should be able to satisfy.