+++ src/horizon/types/offer.ts
@@ -1,5 +1,5 @@
 import { AssetType } from "@stellar/stellar-base";
-import { HorizonApi } from "./../horizon_api";
+import { HorizonApi } from "../horizon_api";
 
 export interface OfferAsset {
   asset_type: AssetType;