+++ test/e2e/src/test-swap.js
@@ -1,183 +1,278 @@
-const test = require("ava");
-const { contract, rpc } = require("../../..");
-const { clientFor, generateFundedKeypair } = require("./util");
+const { expect } = require("chai");
+const { describe, it, before } = require("mocha");
+const {
+  contract,
+  rpc,
+  SorobanDataBuilder,
+  xdr,
+  TransactionBuilder,
+} = require("../../..");
+const {
+  clientFor,
+  generateFundedKeypair,
+  networkPassphrase,
+} = require("./util");
 
 const amountAToSwap = 2n;
 const amountBToSwap = 1n;
 
-test.before(async (t) => {
-  const alice = await generateFundedKeypair();
-  const bob = await generateFundedKeypair();
+describe("Swap Contract Tests", function () {
+  before(async function () {
+    const alice = await generateFundedKeypair();
+    const bob = await generateFundedKeypair();
 
-  const {
-    client: tokenA,
-    contractId: tokenAId,
-    keypair: root,
-  } = await clientFor("token");
-  const { client: tokenB, contractId: tokenBId } = await clientFor("token", {
-    keypair: root,
-  });
-  const { client: swapContractAsRoot, contractId: swapId } = await clientFor(
-    "swap",
-    { keypair: root },
-  );
-  await (
-    await tokenA.initialize({
-      admin: root.publicKey(),
-      decimal: 0,
-      name: "Token A",
-      symbol: "A",
-    })
-  ).signAndSend();
-  await (
-    await tokenA.mint({ amount: amountAToSwap, to: alice.publicKey() })
-  ).signAndSend();
+    const {
+      client: tokenA,
+      contractId: tokenAId,
+      keypair: root,
+    } = await clientFor("token");
+    const { client: tokenB, contractId: tokenBId } = await clientFor("token", {
+      keypair: root,
+    });
+    const { client: swapContractAsRoot, contractId: swapId } = await clientFor(
+      "swap",
+      { keypair: root },
+    );
+    await (
+      await tokenA.initialize({
+        admin: root.publicKey(),
+        decimal: 0,
+        name: "Token A",
+        symbol: "A",
+      })
+    ).signAndSend();
+    await (
+      await tokenA.mint({ amount: amountAToSwap, to: alice.publicKey() })
+    ).signAndSend();
 
-  await (
-    await tokenB.initialize({
-      admin: root.publicKey(),
-      decimal: 0,
-      name: "Token B",
-      symbol: "B",
-    })
-  ).signAndSend();
-  await (
-    await tokenB.mint({ amount: amountBToSwap, to: bob.publicKey() })
-  ).signAndSend();
+    await tokenB
+      .initialize({
+        admin: root.publicKey(),
+        decimal: 0,
+        name: "Token B",
+        symbol: "B",
+      })
+      .then((t) => t.signAndSend());
+    await (
+      await tokenB.mint({ amount: amountBToSwap, to: bob.publicKey() })
+    ).signAndSend();
 
-  t.context = {
-    // eslint-disable-line no-param-reassign
-    root,
-    alice,
-    bob,
-    swapContractAsRoot,
-    swapId,
-    tokenA,
-    tokenAId,
-    tokenB,
-    tokenBId,
-  };
-});
-
-test("calling `signAndSend()` too soon throws descriptive error", async (t) => {
-  const tx = await t.context.swapContractAsRoot.swap({
-    a: t.context.alice.publicKey(),
-    b: t.context.bob.publicKey(),
-    token_a: t.context.tokenAId,
-    token_b: t.context.tokenBId,
-    amount_a: amountAToSwap,
-    min_a_for_b: amountAToSwap,
-    amount_b: amountBToSwap,
-    min_b_for_a: amountBToSwap,
-  });
-  const error = await t.throwsAsync(tx.signAndSend());
-  t.true(
-    error instanceof contract.AssembledTransaction.Errors.NeedsMoreSignatures,
-    `error is not of type 'NeedsMoreSignaturesError'; instead it is of type '${error?.constructor.name}'`,
-  );
-  if (error) t.regex(error.message, /needsNonInvokerSigningBy/);
-});
-
-test("alice swaps bob 10 A for 1 B", async (t) => {
-  const tx = await t.context.swapContractAsRoot.swap({
-    a: t.context.alice.publicKey(),
-    b: t.context.bob.publicKey(),
-    token_a: t.context.tokenAId,
-    token_b: t.context.tokenBId,
-    amount_a: amountAToSwap,
-    min_a_for_b: amountAToSwap,
-    amount_b: amountBToSwap,
-    min_b_for_a: amountBToSwap,
+    this.context = {
+      root,
+      alice,
+      bob,
+      swapContractAsRoot,
+      swapId,
+      tokenA,
+      tokenAId,
+      tokenB,
+      tokenBId,
+    };
   });
 
-  const needsNonInvokerSigningBy = await tx.needsNonInvokerSigningBy();
-  t.is(needsNonInvokerSigningBy.length, 2);
-  t.is(
-    needsNonInvokerSigningBy.indexOf(t.context.alice.publicKey()),
-    0,
-    "needsNonInvokerSigningBy does not have alice's public key!",
-  );
-  t.is(
-    needsNonInvokerSigningBy.indexOf(t.context.bob.publicKey()),
-    1,
-    "needsNonInvokerSigningBy does not have bob's public key!",
-  );
+  it("calling `signAndSend()` too soon throws descriptive error", async function () {
+    const tx = await this.context.swapContractAsRoot.swap({
+      a: this.context.alice.publicKey(),
+      b: this.context.bob.publicKey(),
+      token_a: this.context.tokenAId,
+      token_b: this.context.tokenBId,
+      amount_a: amountAToSwap,
+      min_a_for_b: amountAToSwap,
+      amount_b: amountBToSwap,
+      min_b_for_a: amountBToSwap,
+    });
+    await expect(tx.signAndSend())
+      .to.be.rejectedWith(
+        contract.AssembledTransaction.Errors.NeedsMoreSignatures,
+      )
+      .then((error) => {
+        // Further assertions on the error object
+        expect(error).to.be.instanceOf(
+          contract.AssembledTransaction.Errors.NeedsMoreSignatures,
+          `error is not of type 'NeedsMoreSignaturesError'; instead it is of type '${error?.constructor.name}'`,
+        );
 
-  // root serializes & sends to alice
-  const jsonFromRoot = tx.toJSON();
-  const { client: clientAlice } = await clientFor("swap", {
-    keypair: t.context.alice,
-    contractId: t.context.swapId,
+        if (error) {
+          // Using regex to check the error message
+          expect(error.message).to.match(/needsNonInvokerSigningBy/);
+        }
+      });
   });
-  const txAlice = clientAlice.txFromJSON(jsonFromRoot);
-  await txAlice.signAuthEntries();
 
-  // alice serializes & sends to bob
-  const jsonFromAlice = txAlice.toJSON();
-  const { client: clientBob } = await clientFor("swap", {
-    keypair: t.context.bob,
-    contractId: t.context.swapId,
+  it("modified & re-simulated transactions show updated data", async function () {
+    const tx = await this.context.swapContractAsRoot.swap({
+      a: this.context.alice.publicKey(),
+      b: this.context.bob.publicKey(),
+      token_a: this.context.tokenAId,
+      token_b: this.context.tokenBId,
+      amount_a: amountAToSwap,
+      min_a_for_b: amountAToSwap,
+      amount_b: amountBToSwap,
+      min_b_for_a: amountBToSwap,
+    });
+    await tx.signAuthEntries({
+      publicKey: this.context.alice.publicKey(),
+      ...contract.basicNodeSigner(this.context.alice, networkPassphrase),
+    });
+    await tx.signAuthEntries({
+      publicKey: this.context.bob.publicKey(),
+      ...contract.basicNodeSigner(this.context.bob, networkPassphrase),
+    });
+
+    const originalResourceFee = Number(
+      tx.simulationData.transactionData.resourceFee(),
+    );
+    const bumpedResourceFee = originalResourceFee + 10000;
+
+    tx.raw = TransactionBuilder.cloneFrom(tx.built, {
+      fee: tx.built.fee,
+      sorobanData: new SorobanDataBuilder(
+        tx.simulationData.transactionData.toXDR(),
+      )
+        .setResourceFee(
+          xdr.Int64.fromString(bumpedResourceFee.toString()).toBigInt(),
+        )
+        .build(),
+    });
+
+    await tx.simulate();
+
+    const newSimulatedResourceFee = Number(
+      tx.simulationData.transactionData.resourceFee(),
+    );
+
+    expect(originalResourceFee).to.not.equal(newSimulatedResourceFee);
+    expect(newSimulatedResourceFee).to.be.greaterThan(bumpedResourceFee);
   });
-  const txBob = clientBob.txFromJSON(jsonFromAlice);
-  await txBob.signAuthEntries();
 
-  // bob serializes & sends back to root
-  const jsonFromBob = txBob.toJSON();
-  const { client: clientRoot } = await clientFor("swap", {
-    keypair: t.context.root,
-    contractId: t.context.swapId,
+  it("modified & re-simulated transactions show updated data", async function () {
+    const tx = await this.context.swapContractAsRoot.swap({
+      a: this.context.alice.publicKey(),
+      b: this.context.bob.publicKey(),
+      token_a: this.context.tokenAId,
+      token_b: this.context.tokenBId,
+      amount_a: amountAToSwap,
+      min_a_for_b: amountAToSwap,
+      amount_b: amountBToSwap,
+      min_b_for_a: amountBToSwap,
+    });
+    await tx.signAuthEntries({
+      publicKey: this.context.alice.publicKey(),
+      ...contract.basicNodeSigner(this.context.alice, networkPassphrase),
+    });
+    await tx.signAuthEntries({
+      publicKey: this.context.bob.publicKey(),
+      ...contract.basicNodeSigner(this.context.bob, networkPassphrase),
+    });
+
+    const originalResourceFee = Number(
+      tx.simulationData.transactionData.resourceFee(),
+    );
+    const bumpedResourceFee = originalResourceFee + 10000;
+
+    tx.raw = TransactionBuilder.cloneFrom(tx.built, {
+      fee: tx.built.fee,
+      sorobanData: new SorobanDataBuilder(
+        tx.simulationData.transactionData.toXDR(),
+      )
+        .setResourceFee(
+          xdr.Int64.fromString(bumpedResourceFee.toString()).toBigInt(),
+        )
+        .build(),
+    });
+
+    await tx.simulate();
+
+    const newSimulatedResourceFee = Number(
+      tx.simulationData.transactionData.resourceFee(),
+    );
+
+    expect(originalResourceFee).to.not.equal(newSimulatedResourceFee);
+    expect(newSimulatedResourceFee).to.be.greaterThan(bumpedResourceFee);
   });
-  const txRoot = clientRoot.txFromJSON(jsonFromBob);
 
-  const result = await txRoot.signAndSend();
+  it("alice swaps bob 10 A for 1 B", async function () {
+    const tx = await this.context.swapContractAsRoot.swap({
+      a: this.context.alice.publicKey(),
+      b: this.context.bob.publicKey(),
+      token_a: this.context.tokenAId,
+      token_b: this.context.tokenBId,
+      amount_a: amountAToSwap,
+      min_a_for_b: amountAToSwap,
+      amount_b: amountBToSwap,
+      min_b_for_a: amountBToSwap,
+    });
 
-  t.truthy(
-    result.sendTransactionResponse,
-    `tx failed: ${JSON.stringify(result, null, 2)}`,
-  );
-  t.is(
-    result.sendTransactionResponse.status,
-    "PENDING",
-    `tx failed: ${JSON.stringify(result, null, 2)}`,
-  );
-  t.truthy(
-    result.getTransactionResponseAll?.length,
-    `tx failed: ${JSON.stringify(result.getTransactionResponseAll, null, 2)}`,
-  );
-  t.not(
-    result.getTransactionResponse.status,
-    "FAILED",
-    `tx failed: ${JSON.stringify(
-      result.getTransactionResponse.resultXdr
-        .result()
-        .value()
-        .map((op) => op.value()?.value().switch()),
-      null,
-      2,
-    )}`,
-  );
-  t.is(
-    result.getTransactionResponse.status,
-    rpc.Api.GetTransactionStatus.SUCCESS,
-    `tx failed: ${JSON.stringify(result.getTransactionResponse, null, 2)}`,
-  );
+    const needsNonInvokerSigningBy = await tx.needsNonInvokerSigningBy();
+    expect(needsNonInvokerSigningBy).to.have.lengthOf(2);
+    expect(
+      needsNonInvokerSigningBy.indexOf(this.context.alice.publicKey()),
+    ).to.equal(0, "needsNonInvokerSigningBy does not have alice's public key!");
+    expect(
+      needsNonInvokerSigningBy.indexOf(this.context.bob.publicKey()),
+    ).to.equal(1, "needsNonInvokerSigningBy does not have bob's public key!");
 
-  t.is(
-    (await t.context.tokenA.balance({ id: t.context.alice.publicKey() }))
-      .result,
-    0n,
-  );
-  t.is(
-    (await t.context.tokenB.balance({ id: t.context.alice.publicKey() }))
-      .result,
-    amountBToSwap,
-  );
-  t.is(
-    (await t.context.tokenA.balance({ id: t.context.bob.publicKey() })).result,
-    amountAToSwap,
-  );
-  t.is(
-    (await t.context.tokenB.balance({ id: t.context.bob.publicKey() })).result,
-    0n,
-  );
+    // root serializes & sends to alice
+    const xdrFromRoot = tx.toXDR();
+    const { client: clientAlice } = await clientFor("swap", {
+      keypair: this.context.alice,
+      contractId: this.context.swapId,
+    });
+    const txAlice = clientAlice.txFromXDR(xdrFromRoot);
+    await txAlice.signAuthEntries();
+
+    // alice serializes & sends to bob
+    const xdrFromAlice = txAlice.toXDR();
+    const { client: clientBob } = await clientFor("swap", {
+      keypair: this.context.bob,
+      contractId: this.context.swapId,
+    });
+    const txBob = clientBob.txFromXDR(xdrFromAlice);
+    await txBob.signAuthEntries();
+
+    // bob serializes & sends back to root
+    const xdrFromBob = txBob.toXDR();
+    const { client: clientRoot } = await clientFor("swap", {
+      keypair: this.context.root,
+      contractId: this.context.swapId,
+    });
+    const txRoot = clientRoot.txFromXDR(xdrFromBob);
+
+    await txRoot.simulate();
+    const result = await txRoot.signAndSend();
+
+    expect(result).to.have.property("sendTransactionResponse");
+    expect(result.sendTransactionResponse).to.have.property(
+      "status",
+      "PENDING",
+    );
+    expect(result)
+      .to.have.property("getTransactionResponseAll")
+      .that.is.an("array").that.is.not.empty;
+    expect(result.getTransactionResponse)
+      .to.have.property("status")
+      .that.is.not.equal("FAILED");
+    expect(result.getTransactionResponse).to.have.property(
+      "status",
+      rpc.Api.GetTransactionStatus.SUCCESS,
+    );
+
+    const aliceTokenABalance = await this.context.tokenA.balance({
+      id: this.context.alice.publicKey(),
+    });
+    const aliceTokenBBalance = await this.context.tokenB.balance({
+      id: this.context.alice.publicKey(),
+    });
+    const bobTokenABalance = await this.context.tokenA.balance({
+      id: this.context.bob.publicKey(),
+    });
+    const bobTokenBBalance = await this.context.tokenB.balance({
+      id: this.context.bob.publicKey(),
+    });
+
+    expect(aliceTokenABalance.result).to.equal(0n);
+    expect(aliceTokenBBalance.result).to.equal(amountBToSwap);
+    expect(bobTokenABalance.result).to.equal(amountAToSwap);
+    expect(bobTokenBBalance.result).to.equal(0n);
+  });
 });